Develop

+++ backend/Register/Domain/Repositories/IRequestRepository.cs
@@ -5,6 +5,6 @@ namespace backend.Register.Domain.Repositories;
 public interface IRequestRepository
 {
     Task<IEnumerable<Request>> ListAsync();
-    Task<Request> FindByIdAsync(int id);
+    Task<Request> FindByIdAsync(long id);
     Task AddAsync(Request request);
 }
